backend/requirements.txt
asgiref==3.8.1
Django==4.2.19
django-filter==25.1
djangorestframework==3.16.0
flake8==7.1.2
mccabe==0.7.0
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pyflakes==3.2.0
sqlparse==0.5.3
typing_extensions==4.12.2
